多模块初步化能运行，差so包加载和平台配置化功能

+++ Module_Lib/app/src/main/java/com/ycf/qianzhihe/section/group/activity/GroupSharedFilesActivity.java
@@ -92,10 +92,8 @@ public class GroupSharedFilesActivity extends BaseInitActivity implements OnRefr
     public boolean onContextItemSelected(@NonNull MenuItem item) {
         EaseRecyclerView.RecyclerViewContextMenuInfo info = (EaseRecyclerView.RecyclerViewContextMenuInfo) item.getMenuInfo();
         int position = info.position;
-        switch (item.getItemId()) {
-            case R.id.action_shared_delete :
-                deleteFile(adapter.getItem(position));
-                break;
+        if (item.getItemId() == R.id.action_shared_delete) {
+            deleteFile(adapter.getItem(position));
         }
         return super.onContextItemSelected(item);
     }
